Fix 64bit compatibility issue.
The size of long could be 8 bytes on 64bit env. Change it to
be always 4 bytes on 32/64bit env both.

+++ media/jni/android_media_MediaScanner.cpp
@@ -351,7 +351,7 @@ android_media_MediaScanner_extractAlbumArt(
     if (!data) {
         return NULL;
     }
-    long len = *((long*)data);
+    int len = *((int*)data);
 
     jbyteArray array = env->NewByteArray(len);
     if (array != NULL) {
